Fix broke rails db:* tasks
We need to wrap the collection in a proc so it's not evaluated while
loading this file, but rather is delayed until the associated
ActiveAdmin view is used. Without this, we can't do things like `rails
db:create` b/c this line tries to run, but there's no DB yet!